pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "epub-pipeline"
version = "0.1.0-alpha.1"
description = "Automated pipeline for Ebook metadata enrichment, conversion, and cloud upload."
readme = "README.md"
authors = [
  { name = "Maxime", email = "maxime@example.com" },
]
license = { text = "MIT" }
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "EbookLib>=0.18",
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "termcolor",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "requests-mock>=1.11",
    "ruff",
    "mypy",
    "types-requests",
    "types-termcolor",
]

[project.urls]
Homepage = "https://github.com/your-repo/epub-pipeline"

[project.scripts]
epubpipe = "epub_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["."]  # "." since epub_pipeline is at root

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=epub_pipeline --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]  # E=Error, F=Pyflakes, I=Isort

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.semantic_release]
version_variable = [
    "epub_pipeline/__init__.py:__version__"
]
version_toml = [
    "pyproject.toml:project.version"
]
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "python -m pip install build && python -m build"
